+++ api/routes/products.js
@@ -1,7 +1,8 @@
 const express = require("express");
 const router = express.Router();
-const mongoose = require("mongoose");
 const multer = require('multer');
+const checkAuth = require('../middleware/check-auth');
+const ProductsController = require('../controllers/products');
 
 const storage = multer.diskStorage({
   destination: function(req, file, cb) {
@@ -29,149 +30,14 @@ const upload = multer({
   fileFilter: fileFilter
 });
 
-const Product = require("../models/product");
+router.get("/", ProductsController.products_get_all);
 
-router.get("/", (req, res, next) => {
-  Product.find()
-    .select('name price _id productImage')
-    .exec()
-    .then(docs => {
-       const response = {
-        countDocuments: docs.length,
-        products: docs.map(doc => {
-          return {
-            name: doc.name,
-            price: doc.price,
-            productImage: doc.productImage,
-            _id: doc._id,
-            request: {
-              type: 'GET',
-              url: 'http://localhost:3000/products/' + doc._id
-            }
-          }
-        })
-       }
-      // if (docs.length >= 0) {
-        res.status(200).json(response);
-      // } else {
-      //   res.status(404).json({
-      //     message: 'No entries found'
-      //   });
-      // }
-    })
-    .catch(err => {
-      console.log(err);
-      res.status(500).json({
-        error: err
-      });
-    });
-});
+router.post("/", checkAuth, upload.single('productImage'), ProductsController.products_create_product);
 
-router.post("/", upload.single('productImage'), (req, res, next) => {
-  const product = new Product({
-    _id: new mongoose.Types.ObjectId(),
-    name: req.body.name,
-    price: req.body.price,
-    productImage: req.file.path
-  });
-  product
-    .save()
-    .then(result => {
-      console.log(result);
-      res.status(200).json({
-        message: "Product Created",
-        createdProduct: {
-          name: result.name,
-          price: result.price,
-          _id: result._id,
-          request: {
-            type: 'GET',
-            url: "http://localhost:3000/products/" + result._id
-          }
-        }
-      });
-    })
-    .catch(err => {
-      console.log(err);
-      res.status(500).json({
-        error: err
-      });
-    });
-});
+router.get("/:productId", ProductsController.products_get_product);
 
-router.get("/:productId", (req, res, next) => {
-  const id = req.params.productId;
-  Product.findById(id)
-    .select('name price _id productImage')
-    .exec()
-    .then(doc => {
-      console.log("From database", doc);
-      if (doc) {
-        res.status(200).json({
-            product: doc,
-            request: {
-              type: 'GET',
-              description: "All Products",
-              url: 'http://localhost:3000/products/'
-            }
-        });
-      } else {
-          res.status(404).json({
-            message: "No valid entry found for provided ID"
-          });
-      }
-    })
-    .catch((err) => {
-      console.log(err);
-      res.status(500).json({ error: err });
-    });
-});
+router.patch("/:productId", checkAuth, ProductsController.products_update_product);
 
-router.patch("/:productId", (req, res, next) => {
-  const id = req.params.productId;
-  const updateOps = {};
-  for (const ops of req.body) {
-      updateOps[ops.propName] = ops.value;
-  }
-  Product.updateOne({ _id: id }, { $set: updateOps })
-    .exec()
-    .then(result => {
-      res.status(200).json({
-        message: 'Product Updated',
-        request: {
-          type: 'GET',
-          url: 'http://localhost:3000/products/' + id
-        }
-      });
-    })
-    .catch(err => {
-      console.log(err);
-      res.status(500).json({
-        error: err
-      });
-    });
-  });
-
-router.delete("/:productId", (req, res, next) => {
-  const id = req.params.productId;
-  Product.deleteOne({ _id: id })
-    .exec()
-    .then(result => {
-        res.status(200).json({
-          message: 'Product deleted',
-          request: {
-            type: 'POST',
-            url: 'http://localhost:3000/products/',
-            body: { name: 'String', price: 'Number' }
-          }
-        });
-    })
-    .catch(err => {
-        console.log(err);
-        res.status(500).json({
-          error: err
-        });
-    });
-});
+router.delete("/:productId", checkAuth, ProductsController.products_delete_product);
 
 module.exports = router;
